requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pillow>=10.0.0
numpy>=1.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
azure-storage-blob>=12.19.0
gunicorn>=21.2.0
